Problem: building and installing Postgres extensions
There are certain difficulties with Postgres extension delivery, in
development and production alike, especially when the extensions have
dependencies. They need to be built with the right dependencies for the
platform and delivered in a way that's most suitable for the deployment
environment and preferences.

Solution: design a package manager
that is effectively capturing the while configure-build-[package/install]
cycle, allowing to describe packages declaratively.

It's implemented in Logtalk (Prolog) in order to be able to do
constraint resolution and configuration with the most ease. It is
wrapped into a C applicaiton so it can and should be delivered as a
compiled binary. It has to carry SWI-Prologs .so around, though. But it
is a relatively mild imposition.

This is still an incomplete implementation, as in, it doesn't really
complete the job it's supposed to. More of a sketch at this point,
really.

In broad strokes, it

* allows defining packages in Logtalk
* infers versions from Git, metadata from META.json, make/cc
  requirements from presence of Makefiles and C files
* fetches source code from Git
* is able to find external dependencies using pkg-config
* builds with make

It lacks a few key things:

* finding external dependencies in other ways
* mechanics for building or suggesting how to build external
  dependencies if package managers don't cut it
* reduction requirement satisfactions to a smaller list
* sorting requirement satisfactions by prioritie/preferences
* actually installing packages (or packaging them up)
* recognizing Postgres version
* configuring/building Postgres