Problem: package modules may clash with other objects
If there's ever going to be a package name that clashes with objects in
Logtalk or pgpm itself, it'll try to re-define the object. Definitely
not a great idea.

Solution: generate object names in packages with a counter

Originally, I was re-using package name in the name of the object, but
this makes it tempting to guess the name of the object because we know
how it is constructed. With a counter we can't really do this, so we
can avoid this error-prone approach.